add upload, post and get for raports
